Pass an `apis` object to the user commands.
This object provides the users with a fluent API to use to implement
their commands.  The API provides utilities to access components from
the environ, as well as to use the MarkLogic APIs, to inspect the
files in the project, and to issue any HTTP requests.

The HTTP requests in particular (specific API invocations or any HTTP
request) use the information from the environ (connection infos, host,
credentials, API connection infos, ports, endpoints, etc.)

If you override something in your environs, or simply if you use a
different environ with a different host and/or different credentials,
your command will automatically use them, through the apis object.