Make the platform HTTP functions more generic.
The get, post and put functions on the platform object are more
generic now.  Checking the response codes and using the response
content is the responsibility of the actions using them.

They are then reusable by the apis object in the user commands, to
issue more general-purposes HTTP calls.

Also, ensure we display action message even on dry mode.