Walking imports must be done always in the same way.
- higher precedence on the right-most import
- resolving params and resolving components must use the same precedence
- so reverse the import array, and walk alway forward through it

The use case is to "attach" extra features to an existing, most-likely
full fledged environ like this:

    mlproj -e main/extra show

This is equivalent to `imports: ['main', 'extra']`.  Anything found in
`extra` overrides things from `main`.

Usually, if you have a `base` environ that all your environs inherit
from (with a basic version of your databases for instance), `extra`
should not import it.  If it does, the values from `base`, if not
overriden in `extra`, will have precedence over `main` (*via* `extra`.)